refactor: renamed legacy table component with Legacy prefix
