+++ apps/main/src/llamalend/features/market-list/drawers/MarketListFilterDrawer.tsx
@@ -1,10 +1,11 @@
 import { LlamaMarket } from '@/llamalend/queries/market-list/llama-markets'
-import { Button, Grid } from '@mui/material'
+import { Grid } from '@mui/material'
 import { useSwitch } from '@ui-kit/hooks/useSwitch'
 import { t } from '@ui-kit/lib/i18n'
 import { FilterIcon } from '@ui-kit/shared/icons/FilterIcon'
 import { FilterProps } from '@ui-kit/shared/ui/DataTable/data-table.utils'
 import { HiddenCountResetButton } from '@ui-kit/shared/ui/DataTable/HiddenCountResetButton'
+import { SelectableChip } from '@ui-kit/shared/ui/SelectableChip'
 import { DrawerHeader } from '@ui-kit/shared/ui/SwipeableDrawer/DrawerHeader'
 import { DrawerItems } from '@ui-kit/shared/ui/SwipeableDrawer/DrawerItems'
 import { SwipeableDrawer } from '@ui-kit/shared/ui/SwipeableDrawer/SwipeableDrawer'
@@ -33,22 +34,20 @@ export const MarketListFilterDrawer = ({
   userPositionsTab,
   ...filterProps
 }: Props) => {
-  const [open, openDrawer, closeDrawer] = useSwitch(false)
+  const [open, , closeDrawer, toggleDrawer] = useSwitch(false)
   const hasPopularFilters = userPositionsTab === MarketRateType.Borrow || !userPositionsTab
   const showUserChips = !userPositionsTab
   return (
     <SwipeableDrawer
       paperSx={{ maxHeight: SizesAndSpaces.MaxHeight.drawer }}
       button={
-        <Button
-          variant="outlined"
-          size="small"
-          fullWidth
-          onClick={openDrawer}
+        <SelectableChip
+          size="medium"
+          selected={open}
+          icon={<FilterIcon />}
+          toggle={toggleDrawer}
           data-testid="btn-drawer-filter-lamalend-markets"
-        >
-          {t`Filter`} {hiddenCount ? `(${hiddenCount})` : ''} <FilterIcon sx={{ marginLeft: Spacing.sm }} />
-        </Button>
+        />
       }
       open={open}
       setOpen={closeDrawer}
