+++ apps/main/src/llamalend/features/user-position-history/UserEventsTable.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react'
 import { SortingState } from '@tanstack/react-table'
 import { t } from '@ui-kit/lib/i18n'
 import { getTableOptions, useTable } from '@ui-kit/shared/ui/DataTable/data-table.utils'
-import { DataTable } from '@ui-kit/shared/ui/DataTable/DataTable'
 import { EmptyStateRow } from '@ui-kit/shared/ui/DataTable/EmptyStateRow'
+import { LegacyDataTable } from '@ui-kit/shared/ui/DataTable/LegacyDataTable'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
 import { DEFAULT_SORT, USER_POSITION_HISTORY_COLUMNS } from './columns'
 import { ParsedUserCollateralEvent } from './hooks/useUserCollateralEvents'
@@ -34,7 +34,7 @@ export const UserEventsTable = ({ events, loading, isError }: UserEventsTablePro
   })
 
   return (
-    <DataTable
+    <LegacyDataTable
       table={table}
       emptyState={
         <EmptyStateRow table={table}>{isError ? t`Could not load events` : t`No events found`}</EmptyStateRow>
