+++ packages/curve-ui-kit/src/shared/ui/DataTable/LegacyDataTable.tsx
@@ -16,7 +16,7 @@ import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
 import { type TableItem, type TanstackTable } from './data-table.utils'
 import { DataRow, type DataRowProps } from './DataRow'
 import { HeaderCell } from './HeaderCell'
-import { NewFilterRow } from './NewFilterRow'
+import { LegacyFilterRow } from './LegacyFilterRow'
 import { SkeletonRows } from './SkeletonRows'
 import { TableViewAllCell } from './TableViewAllCell'
 import { useTableRowLimit } from './useTableRowLimit'
@@ -69,7 +69,7 @@ const { Sizing } = SizesAndSpaces
 /**
  * DataTable component to render the table with headers and rows.
  */
-export const NewDataTable = <T extends TableItem>({
+export const LegacyDataTable = <T extends TableItem>({
   emptyState,
   children,
   loading,
@@ -111,6 +111,7 @@ export const NewDataTable = <T extends TableItem>({
     position: 'sticky',
     top: maxHeight ? 0 : top,
     zIndex: t.zIndex.tableHeader,
+    backgroundColor: t.design.Table.Header.Fill,
     marginBlock: Sizing.sm,
   })
   const showFooter = showPagination || showViewAllButton || footerRow
@@ -119,13 +120,14 @@ export const NewDataTable = <T extends TableItem>({
     <WithWrapper Wrapper={Box} shouldWrap={maxHeight} sx={{ maxHeight, overflowY: 'auto' }} ref={containerRef}>
       <Table
         sx={{
+          backgroundColor: t => t.design.Layer[1].Fill,
           borderCollapse: 'separate' /* Don't collapse to avoid funky stuff with the sticky header */,
         }}
         data-testid={!loading && 'data-table'}
       >
         {!hideHeader && (
           <TableHead sx={tableHeaderSx} data-testid="data-table-head">
-            {children && <NewFilterRow table={table}>{children}</NewFilterRow>}
+            {children && <LegacyFilterRow table={table}>{children}</LegacyFilterRow>}
 
             {headerGroups.map(headerGroup => (
               <TableRow key={headerGroup.id} sx={{ height: Sizing.xxl }}>
@@ -160,7 +162,7 @@ export const NewDataTable = <T extends TableItem>({
         </TableBody>
         {showFooter && (
           <TableFooter>
-            {footerRow && <TableRow>{footerRow}</TableRow>}
+            {footerRow && <TableRow sx={{ verticalAlign: rowProps.verticalAlign }}>{footerRow}</TableRow>}
             {showViewAllButton && (
               <TableRow>
                 <TableViewAllCell colSpan={columnCount} onClick={onShowAll} isLoading={isLoadingViewAll}>
