+++ packages/curve-ui-kit/src/shared/ui/DataTable/LegacyTableFiltersTitles.tsx
@@ -1,7 +1,7 @@
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
 
-export const TableFiltersTitles = ({ title, subtitle }: { title: string; subtitle?: string }) => (
+export const LegacyTableFiltersTitles = ({ title, subtitle }: { title: string; subtitle?: string }) => (
   <Stack height="100%" justifyContent="end">
     <Typography variant="headingSBold">{title}</Typography>
     {subtitle && (
