+++ packages/curve-ui-kit/src/shared/ui/DataTable/TableFilters.tsx
@@ -1,17 +1,13 @@
 import { ReactNode, useRef } from 'react'
+import Box from '@mui/material/Box'
 import Collapse from '@mui/material/Collapse'
-import Fade from '@mui/material/Fade'
 import Grid from '@mui/material/Grid'
 import Stack from '@mui/material/Stack'
 import { useIsMobile } from '@ui-kit/hooks/useBreakpoints'
 import { useDebounce } from '@ui-kit/hooks/useDebounce'
-import { useFilterExpanded } from '@ui-kit/hooks/useLocalStorage'
 import { useSwitch } from '@ui-kit/hooks/useSwitch'
-import { Duration } from '@ui-kit/themes/design/0_primitives'
+import { GearIcon } from '@ui-kit/shared/icons/GearIcon'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { FilterIcon } from '../../icons/FilterIcon'
-import { GearIcon } from '../../icons/GearIcon'
-import { ReloadIcon } from '../../icons/ReloadIcon'
 import { TableButton } from './TableButton'
 import { TableSearchField } from './TableSearchField'
 import { TableVisibilitySettingsPopover } from './TableVisibilitySettingsPopover'
@@ -23,69 +19,66 @@ const { Spacing } = SizesAndSpaces
  * A component that wraps a table and provides a title, subtitle, and filter controls.
  */
 export const TableFilters = <ColumnIds extends string>({
-  leftChildren,
+  header,
   filterExpandedKey,
-  loading,
-  onReload,
+  filterExpanded,
   visibilityGroups,
   toggleVisibility,
-  hasSearchBar,
   collapsible,
   chips,
+  filterChip,
+  sortChip,
   searchText,
   disableSearchAutoFocus,
   onSearch,
 }: {
-  leftChildren: ReactNode
+  header: ReactNode
   filterExpandedKey: string
-  loading: boolean
-  onReload?: () => void
+  filterExpanded: boolean
   visibilityGroups: VisibilityGroup<ColumnIds>[]
   toggleVisibility?: (columns: string[]) => void
-  hasSearchBar?: boolean
   collapsible?: ReactNode // filters that may be collapsed
   chips?: ReactNode // buttons that are part of the collapsible (on mobile) or always visible (on larger screens)
+  filterChip?: ReactNode // buttons responsible for filtering
+  sortChip?: ReactNode // buttons responsible for sorting
   searchText: string // text to search for, only used for mobile
   disableSearchAutoFocus?: boolean
   onSearch: (value: string) => void
 }) => {
-  const [filterExpanded, setFilterExpanded] = useFilterExpanded(filterExpandedKey)
   const [visibilitySettingsOpen, openVisibilitySettings, closeVisibilitySettings] = useSwitch()
   const settingsRef = useRef<HTMLButtonElement>(null)
   // search is here because we remove the table title when searching on mobile
   const isMobile = useIsMobile()
-  const [isSearchExpanded, , , toggleSearchExpanded] = useSwitch(!isMobile)
   const [searchValue, setSearchValue] = useDebounce({ initialValue: searchText, callback: onSearch })
   const isCollapsible = collapsible || (isMobile && chips)
-  const isExpandedOrValue = Boolean(isSearchExpanded || searchValue)
-  const hideTitle = hasSearchBar && isExpandedOrValue && isMobile
 
   return (
-    <Stack paddingBlockEnd={{ mobile: Spacing.sm.tablet }} paddingBlockStart={{ mobile: Spacing.md.tablet }}>
-      <Grid container spacing={Spacing.sm} paddingInline={Spacing.md} justifyContent="space-between">
-        <Fade in={!hideTitle} timeout={Duration.Transition} mountOnEnter unmountOnExit>
-          <Grid size={{ mobile: 'grow', tablet: 6 }} sx={{ position: hideTitle ? 'absolute' : 'relative' }}>
-            {leftChildren}
-          </Grid>
-        </Fade>
+    <Stack sx={{ backgroundColor: t => t.design.Layer[1].Fill }}>
+      {header}
+      <Grid container spacing={Spacing.lg} padding={Spacing.sm} justifyContent="space-between" alignItems="center">
         <Grid
-          size={{ mobile: isExpandedOrValue ? 12 : 'auto', tablet: 6 }}
-          display="flex"
-          justifyContent="flex-end"
-          gap={Spacing.xs}
-          flexWrap="wrap"
+          size={{ mobile: 12, tablet: 7 }}
+          sx={{
+            display: 'flex',
+            // overlap adjacent component to avoid duplicate border width
+            '& > .tableControl + .tableControl': { ml: '-1px' },
+          }}
         >
-          {hasSearchBar && (
+          {/* Box wrapper needed for applying style */}
+          {filterChip && <Box className="tableControl">{filterChip}</Box>}
+          <Box className="tableControl" sx={{ flex: 1, minWidth: 0 }}>
             <TableSearchField
               value={searchValue}
               onChange={setSearchValue}
               testId={filterExpandedKey}
-              {...(isMobile && { toggleExpanded: toggleSearchExpanded })}
-              isExpanded={isExpandedOrValue}
               disableAutoFocus={disableSearchAutoFocus}
             />
-          )}
-          {!isMobile && toggleVisibility && (
+          </Box>
+          {sortChip && <Box className="tableControl">{sortChip}</Box>}
+        </Grid>
+        {!isMobile && (
+          <Grid container size="grow" spacing="none" justifyContent="flex-end">
+            {chips}
             <TableButton
               ref={settingsRef}
               onClick={openVisibilitySettings}
@@ -93,20 +86,8 @@ export const TableFilters = <ColumnIds extends string>({
               testId="btn-visibility-settings"
               active={visibilitySettingsOpen}
             />
-          )}
-          {isCollapsible && !isMobile && (
-            <TableButton
-              onClick={() => setFilterExpanded(prev => !prev)}
-              active={filterExpanded}
-              icon={FilterIcon}
-              testId="btn-expand-filters"
-            />
-          )}
-          {onReload && !isMobile && <TableButton onClick={onReload} icon={ReloadIcon} rotateIcon={loading} />}
-        </Grid>
-        <Grid container size={12} spacing={Spacing.sm} justifyContent="space-between">
-          {chips}
-        </Grid>
+          </Grid>
+        )}
       </Grid>
       {isCollapsible && !isMobile && <Collapse in={filterExpanded}>{collapsible}</Collapse>}
